requirements.txt
cookiecutter
attrs
pathlib_mate
